docs: Update docs to simplify how we make changes via Github.
In a lot of cases, github will now automatically make forks and branchs so I
removed a bunch of related steps. They just need to be looking at our root repo
and if they don't have permission to edit it, github will automatically fork it
for them and create a branch.