docs: Cleanup top level of the source folder.
Remove some more items from the top level of the source folder.  The
`support_learners` section was empty and not clear what persona that was
supposed to be for so I've remove it for now.

I've also moved the feedback form to an `other` space.  I'm already
thinking of a general "getting help" page that could also live under
other.