build!: Prefix section labels with the document path.
This is a fairly major change since it changes the scope of how `:ref:`
links work.  Before this change, any section header would get a label
created for it that you could reference globally.  After this change,
ref links will need to be updated to contain the document the section is
in.

The old way worked fine when all your sections were likely to have unique names
but with a lot of our reference and how-to documentation this is not the
case.  They will often have the same subsection names.  For example, all
the decision records(ADRs) follow a template and so will all have a
`Status`, `Context` and `Decision` section.  This breaks the build
because you can't have multiple sections with the same link label.

Downsides of the change:

    * We need to update any existing links.
    * Links to subsections could break if the subsection is moved to a
      different file.
        * We'll catch this in CI so I'm not too worried about it.