build: use canonical upgrade-python workflow
Workflow scripts are defined in the .github repo under
workflow-templates. This commit copies the changes from that
workflow into this repo.
In addition to bringing us in line with the default,
this fixes an issue where the workflow was trying
to run with the branch name `"main"` instead of
`main`, resulting in the opaque error of:

    The process '/usr/bin/git' failed with exit code 1