docs: Remove the under development banner.
We are now done with the initial development and there is enough content
that we feel comfortable pointing people to this site as our current
docs home page.  We have plans to expand the content but we don't need
to wait for all of that to remove the banner.

Also, correct an out of date link to openedx.org