docs: Add a community card on the main index.
To do this we're also removing the what's new section which can be a bit
confusing since it has no content.  It also has led to some confusion
where people assume it leads to the given persona section but it does
not.

Eventually the community section will not link out to other places but
just have the documents in this project but cross link for now so that
we point to the current canonical content.