FC-0068 docs: Reference cleaning by edunext #2
This PR modifies a series of reference documents migrated from the legacy repositories to the educators' documentation.

The principal changes are in the custom_python reference, divided into one reference, two concepts, and three how-tos.

Another important thing for reviewers is the link to edX files in the Custom Javascript reference that I couldn't erase due to their need for the proper function of that example.