FC-0068: Reference cleaning by edunext #4
This PR modifies a series of reference documents migrated from the legacy repositories to the educators' documentation.
The principal changes in the lti_canvas_example and lti_blackboard_example were converted into how-tos. An issue will be created for each document for an extended review.

In the case of drag_and_drop and numerical_input, there is one reference to edX legacy repositories in each document that I could not erase or replace with another link.