FC-0068: Reference cleaning by edunext #5
This PR modifies a series of reference documents migrated from the legacy repositories to the educators' documentation.
The principal changes were in the erased checkbox document, and the Oppia exploration document was divided into a reference and a how-to.

Also, the document create_exercises_and_tools has a link to an edX repository file about Levels Support that is not in the Open edX repository.