FC-0068 docs: Reference cleaning by edunext #6
This PR modifies a series of reference documents migrated from the legacy repositories to the educators' documentation.

The principal changes are in the Access_ORA_Info, divided into 5 how-tos and a reference.
Also, I added some content in the multiple_choice to adapt that page to the parameters defined in the Style Guide.