feat:  adding documentation for  course certificate templates
The existing  documentation on custom certificate templates within an
instance is actually enough to make a custom certificate for something
other than languages, but doesn't actually make it clear that's what
it's doing.  This restructures the documentation to make it more obvious
that the  custom certificate templates work for levels other than
language.

(For an example of people asking about this, see [this discourse
thread](https://discuss.openedx.org/t/trouble-customizing-course-certificates-on-open-edx-tutor-v19/16010)
in which people ask for documentation but there's nothing obvious to link
them to.)